Feat: show an ellipsis when there are more than 6 pages in Common/Pagination
